Move depsdev to internal/collector
